app backend: create feed_forward function
