+++ src/utils/index.ts
@@ -8,3 +8,4 @@ export * from "./generate-integer";
 export * from "./hex-to-rgb";
 export * from "./rgb-to-hex";
 export * from "./set-input-defaults";
+export * from "./get-input-for-nn";
